requirements.txt
nanocamera
Jetson.GPIO